Eliminate the performance impact of operands in Array#union and Array#intersection.
When I do intersection (`a & b`) or union (` a | b`) usualy one of the arrays is
more likely longer then the second one. I always try to remember in what order
should I write `a` and` b` for the best performance.

Here I suggest to make it simpler and eliminate the difference.